+++ lib/pubnub/client/events.rb
@@ -6,6 +6,7 @@ module Pubnub
     module Events
       EVENTS = %w[publish subscribe presence leave history here_now audit grant grant_token revoke_token delete_messages
                   revoke time heartbeat where_now set_state state channel_registration message_counts signal
+                  add_message_action get_message_actions remove_message_action
                   add_channels_to_push list_push_provisions remove_channels_from_push remove_device_from_push
                   set_uuid_metadata set_channel_metadata remove_uuid_metadata remove_channel_metadata
                   get_uuid_metadata get_all_uuid_metadata get_channel_metadata get_all_channels_metadata
@@ -15,11 +16,13 @@ module Pubnub
       EVENTS.each do |event_name|
         define_method event_name do |options = {}, &block|
           options[:callback] = block if options[:callback].nil?
+          # Use constructor-provided :callback if nothing passed to method call.
+          options[:callback] = self.env[:callback] if options[:callback].nil?
           event = Pubnub.const_get(
             Formatter.classify_method(event_name)
           ).new(options, self)
 
-          if options[:http_sync]
+          if event.sync?
             event.fire
           elsif event.is_a? SubscribeEvent
             @subscriber.add_subscription(event)