+++ spec/examples/revoke_examples_1_spec.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require 'helpers/spec_helper'
 
 describe Pubnub::Revoke do
   around :each do |example|