+++ spec/examples/status_request_message_count_exceeded_spec.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require 'helpers/spec_helper'
 
 describe "Status request message count exceeded" do
   around :each do |example|