+++ spec/lib/events/leave_spec.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require 'helpers/spec_helper'
 
 describe Pubnub::Leave do
   it_behaves_like "an event"