+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+## v5.2.1
+December 13 2022
+
+#### Fixed
+- Fix format which is used to return server data from `get_message_actions` method call.
+
+## v5.2.0
+December 12 2022
+
+#### Added
+- Add api which allow to work with message actions endpoints.
+
+#### Modified
+- Migrate test suite run from Travis to GitHub Actions.
+
+## v5.1.2
+November 23 2022
+
+#### Fixed
+- Fix issue because of which channel hasn't been encoded properly with endpoint call.
+
+## v5.1.1
+October 26 2022
+
+#### Fixed
+- Fix issue because of which `callback` and `http_sync` provided during client configuration not used when missing in method call.
+
 ## v5.1.0
 July 26 2022